requirements.txt
aiosqlite==0.17.0
alembic==1.8.1
anyio==3.6.2
apprise==1.2.0
asgi-lifespan==2.0.0
asyncpg==0.27.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
commonmark==0.9.1
coolname==2.1.0
croniter==1.3.8
cryptography==38.0.4
docker==6.0.1
fastapi==0.88.0
fsspec==2022.11.0
google-auth==2.15.0
greenlet==2.0.1
griffe==0.24.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.2
httpx==0.23.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==5.1.0
jsonpatch==1.32
jsonpointer==2.3
kubernetes==25.3.0
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
oauthlib==3.2.2
orjson==3.8.3
packaging==22.0
pathspec==0.10.2
pendulum==2.1.2
prefect==2.7.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.6
pytzdata==2020.1
PyYAML==6.0
readchar==4.0.3
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.22.0
text-unidecode==1.3
toml==0.10.2
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
websocket-client==1.4.2
zipp==3.11.0
